chanege the path of backstage file
