+++ php/backstage/config.php
@@ -1,0 +1,27 @@
+<?php
+/* Database credentials. Assuming you are running MySQL
+server with default setting (user 'root' with no password) */
+/*define() 函數定義一個常量。
+在設定以後，常量的值無法更改
+常量名不需要開頭的美元符號 ($)
+作用域不影響對常量的訪問
+常量值只能是字符串或數字*/
+define('DB_SERVER', 'localhost');
+define('DB_USERNAME', 'root');
+define('DB_PASSWORD', '94keep30678');
+define('DB_NAME', 'parkinglot');
+ 
+/* Attempt to connect to MySQL database */
+$link = new mysqli(DB_SERVER, DB_USERNAME, DB_PASSWORD, DB_NAME);
+// 輸入中文也OK的編碼
+mysqli_query($link, 'SET NAMES utf8mb4');
+
+// Check connection
+if($link === false){
+    die("ERROR: Could not connect. " . mysqli_connect_error());
+}
+else{
+    return $link;
+}
+
+?>
